% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wd_above.r
\name{set_wd_to_above}
\alias{set_wd_to_above}
\title{Set working directory to one level above the current file's directory}
\usage{
set_wd_to_above()
}
\value{
Sets the working directory
}
\description{
This function sets the working directory to the parent directory of the directory
where the currently active RStudio file is located.
}
