% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wd_to_current.r
\name{set_wd_to_current}
\alias{set_wd_to_current}
\title{Set working directory to the current file's directory}
\usage{
set_wd_to_current()
}
\value{
Sets the working directory
}
\description{
This function sets the working directory to the directory where the currently active RStudio file is located.
}
